+++ src/vulkan/device_allocator/allocation.rs
@@ -1,6 +1,13 @@
 use ash::vk;
 
-use super::Allocation;
+/// A single allocated piece of device memory.
+#[derive(Clone, Debug, Copy, PartialEq, Eq)]
+pub struct Allocation {
+    pub memory: vk::DeviceMemory,
+    pub offset: vk::DeviceSize,
+    pub byte_size: vk::DeviceSize,
+    pub memory_type_index: u32,
+}
 
 impl Allocation {
     /// An empty allocation with null pointers and resource references.
