+++ src/vulkan/pipeline/pipeline_layout.rs
@@ -2,8 +2,16 @@ use std::sync::Arc;
 
 use ash::{version::DeviceV1_0, vk};
 
-use super::{PipelineError, PipelineLayout};
-use crate::vulkan::{errors::VulkanDebugError, RenderDevice, VulkanDebug};
+use crate::vulkan::{
+    errors::VulkanDebugError, pipeline::PipelineError, RenderDevice,
+    VulkanDebug,
+};
+
+/// An owned Pipeline Layout which is destroyed automatically when it's dropped.
+pub struct PipelineLayout {
+    pub raw: vk::PipelineLayout,
+    pub vk_dev: Arc<RenderDevice>,
+}
 
 impl PipelineLayout {
     pub fn new(
