+++ src/vulkan/render_device/gpu_queue.rs
@@ -1,6 +1,12 @@
 use ash::vk;
 
-use super::GpuQueue;
+/// This struct bundles a Vulkan queue with related data for easy tracking.
+#[derive(Debug, Clone, Copy)]
+pub struct GpuQueue {
+    pub queue: vk::Queue,
+    pub family_id: u32,
+    pub index: u32,
+}
 
 impl GpuQueue {
     /// Build a queue wrapper from the raw queue handle.
