+++ src/vulkan/render_pass/mod.rs
@@ -1,17 +1,4 @@
 mod render_pass;
+mod render_pass_error;
 
-use ::{ash::vk, std::sync::Arc, thiserror::Error};
-
-use crate::vulkan::RenderDevice;
-
-#[derive(Debug, Error)]
-pub enum RenderPassError {
-    #[error("Unable to create a new render pass")]
-    UnableToCreateRenderPass(#[source] vk::Result),
-}
-
-/// An owned Vulkan RenderPass which automatically destroys itself when dropped.
-pub struct RenderPass {
-    pub raw: vk::RenderPass,
-    pub vk_dev: Arc<RenderDevice>,
-}
+pub use self::{render_pass::RenderPass, render_pass_error::RenderPassError};
