+++ src/vulkan/render_pass/render_pass_error.rs
@@ -1,0 +1,7 @@
+use ::{ash::vk, thiserror::Error};
+
+#[derive(Debug, Error)]
+pub enum RenderPassError {
+    #[error("Unable to create a new render pass")]
+    UnableToCreateRenderPass(#[source] vk::Result),
+}
