+++ src/vulkan_ext/command_buffer_ext_error.rs
@@ -1,0 +1,12 @@
+use ::{ash::vk, thiserror::Error};
+
+#[derive(Debug, Error)]
+pub enum CommandBufferExtError {
+    #[error("Unable to begin the command buffer")]
+    UnableToBeginCommandBuffer(#[source] vk::Result),
+
+    #[error("Unable to end the command buffer")]
+    UnableToEndCommandBuffer(#[source] vk::Result),
+}
+
+pub type CommandResult<T> = Result<T, CommandBufferExtError>;
