+++ src/vulkan_ext/mod.rs
@@ -1,3 +1,7 @@
 mod command_buffer_ext;
+mod command_buffer_ext_error;
 
-pub use self::command_buffer_ext::CommandBufferExt;
+pub use self::{
+    command_buffer_ext::CommandBufferExt,
+    command_buffer_ext_error::{CommandBufferExtError, CommandResult},
+};
